Move pkg/legacy to pkg/output/legacy
