+++ pkg/output/legacy/create-component-entity.go
