+++ pkg/output/legacy/create-group-entity.go
