+++ pkg/output/legacy/create-user-entity.go
